+++ inc/ti/task.h
@@ -26,6 +26,7 @@ ti_task_t * ti_task_create(uint64_t change_id, ti_thing_t * thing);
 ti_task_t * ti_task_get_task(ti_change_t * change, ti_thing_t * thing);
 void ti_task_destroy(ti_task_t * task);
 int ti_task_add_set_add(ti_task_t * task, ti_raw_t * key, vec_t * added);
+int ti_task_add_set_clear(ti_task_t * task, ti_raw_t * key);
 int ti_task_add_set(ti_task_t * task, ti_raw_t * key, ti_val_t * val);
 int ti_task_add_new_type(ti_task_t * task, ti_type_t * type);
 int ti_task_add_to_type(ti_task_t * task, ti_type_t * type);